tests/wpt/tests/html/canvas/offscreen/line-styles/2d.line.miter.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.line.miter.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.miter.invalid</h1>
<p class="desc">Setting miterLimit to invalid values is ignored</p>


<script>
var t = async_test("Setting miterLimit to invalid values is ignored");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.miterLimit = 1.5;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = 0;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = -1;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = Infinity;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = -Infinity;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = NaN;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = 'string';
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = true;
  _assertSame(ctx.miterLimit, 1, "ctx.miterLimit", "1");

  ctx.miterLimit = 1.5;
  ctx.miterLimit = false;
  _assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");
  t.done();

});
</script>
